Add team member while creating

+++ backend/EasyMeets.Core/EasyMeets.Core.BLL/Interfaces/ITeamService.cs
@@ -8,5 +8,5 @@ public interface ITeamService
     Task<bool> ValidatePageLinkAsync(long teamId, string pageLink);
     Task<TeamDto> CreateTeamAsync(NewTeamDto teamDto);
     Task UpdateTeamAsync(TeamDto team);
-    Task DeleteTeamAsync(int teamId);
+    Task DeleteTeamAsync(long teamId);
 }
